# openapi.yaml
openapi: 3.1.0
info: 
  title: ticket booking system
  version: 1.0.0
  description: documentation API using yaml
security:
  - bearerAuth: []

servers: 
  - url: http://localhost:3001
    description: local development

components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        email:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    RegisterRequest:
      type: object
      properties:
        username: 
          type: string
        name:
          type: string
        email:
          type: string
        password:
          type: string
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/User'
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
    Event:
      type: object
      properties:
        id:
          type: number
        name:
         type: string
        event_date:
          type: string
          format: date-time
        description:
          type: string
        address:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateEventRequest:
      type: object
      properties:
        name:
          type: string
        event_date:
          type: string
          format: date-time
        description:
          type: string
        address:
          type: string
    Ticket:
      type: object
      properties:
        id:
          type: number
        event_id:
          type: number
        name:
          type: string
        available_quota:
          type: integer
        total_quota:
          type: integer
        price:
          type: number
    CreateTicketRequest:
      type: object
      properties:
        name:
          type: string
        total_quota:
          type: integer
        price:
          type: number
    OrderDetail:
      type: object
      properties:
        id:
          type: number
        ticket_id:
          type: number
        price:
          type: number
        quantity:
          type: integer
        subtotal:
          type: number
    Order:
      type: object
      properties:
        id:
          type: number
        user_id:
          type: number
        total_amount:
          type: number
        status:
          type: string
          enum:
          - PENDING
          - SUCCESS
          - EXPIRED
          - CANCELLED 
        created_at:
          type: string
          format: date-time
        expired_at:
          type: string
          format: date-time
        idempotency_key:
          type: string
          format: uuid 
        details:
          type: array
          items:
            $ref: '#/components/schemas/OrderDetail'
    CreateOrderRequest:
      type: object
      properties:
        items:                 
          type: array
          items:               
            type: object
            properties:
              ticket_id:
                type: integer
              quantity:
                type: integer
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

paths: 
  /auth/register:
    post:
      operationId: registerUser
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses: 
        '201':
          description: Created
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      operationId: loginUser
      summary: User Login Request
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Success
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /events:
    get:
      operationId: getEvents
      summary: Get all events
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: createEvent
      summary: Create a new event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /events/{id}:
    parameters:
    - name: id
      in: path
      required: true 
      schema: 
        type: integer
    get:
      operationId: getEventById
      summary: Get event detail by ID
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      operationId: updateEventById
      summary: Update an event by id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/Event'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      operationId: deleteEventById
      summary: Delete an event by id
      responses:
        '200':
          description: Event deleted successfully
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /events/{id}/tickets:
    parameters:
    - name: id
      in: path
      required: true 
      schema: 
        type: integer
    get:
      operationId: getTickets
      summary: Get all tickets for a specific event
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ticket'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: createTicket
      summary: Create a new ticket for an event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTicketRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'        
  /order:
    post: 
      operationId: createOrder
      summary: Create an order
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        description: Unique key to prevent double booking
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse' 
    get:
      operationId: getOrderHistory
      summary: Get user order history
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /order/{id}/confirm:
    parameters:
    - name: id
      in: path
      required: true 
      schema: 
        type: integer
    post:
      operationId: confirmPayment
      summary: Confirm order payment
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'      